Dummy Bike Programme Data

A dummy data set documenting a two stage level intervention programme of the school scenario in the vignette. Report ...

Usage

its_data_bike_programme

Format

## `its_data_bike_programme` A data frame with 104 rows and 4 columns:

Date

Weekly dates

group_var

Variable to indicate either treatment or control row of interest

Period

Description of intervention phase

score

The outcome metric of the scenario - weekly total maintenance incidents"

...